Get MKDAKernel, KDAKernel, and ALEKernel working
- Switch from nltools’ create_sphere to using adapted versions of
Neurosynth’s code for MKDAKernel and KDAKernel
- Move ALE kernel utilities from nimare.utils to inmate.meta.cbma.utils
- Remove nltools from list of dependencies
- Add ijk to Dataset.coordinates (pandas data frame of contrasts’
coordinates and related values)
