+++ nimare/version.py
@@ -60,4 +60,4 @@ MAJOR = _version_major
 MINOR = _version_minor
 MICRO = _version_micro
 VERSION = __version__
-REQUIRES = ["nibabel", "numpy", "scipy", "pandas", "nltools"]
+REQUIRES = ["nibabel", "numpy", "scipy", "pandas"]
